refactor: Input 공통 스타일 분리 및 디렉토리 구조 변경
