+++ frontend/src/stories/DateInput.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import DateTimeInput from '../components/@common/dateTimeInput/DateTimeInput';
+import DateTimeInput from '../components/@common/inputs/DateTimeInput';
 
 const meta: Meta<typeof DateTimeInput> = {
   title: 'Components/Input/Date',
